Change type to uint32_t.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1518 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ swap.h
@@ -1,6 +1,8 @@
 #ifndef SWAP_H
 #define SWAP_H
 
-long swap_32bit(long x);
+#include <inttypes.h>
+
+uint32_t swap_32bit(uint32_t x);
 
 #endif
